HTML editing and finishing CSS
Edited the HTML to have more of the required tags and added more
dialog. For the css, I edited the tags based around the changes on
HTML. Finished the Media query for 60em and 90em, all contents are now
resized to their respective query. Also edited the 25em and 38em query
based on the changes to the html